socket.io impemented for auction page
