+++ server/server.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const app = express();
+
+const server = require('http').createServer(app);
+const io = require('socket.io')(server, {
+  cors: { origin: 'http://localhost:3000' },
+});
+module.exports = { server, io, app };
